Reformat all java files
Due to problems with the CatrobatBot format job, the behavior has been
changed. Instead of checking for unformatted code only in the VCS
changes, all java source code will be checked for now on. Checkstyle,
PMD and Lint also work this way.

The initial step requieres a reformat of the whole project.