Fail release build if Crashlytics credentials are missing
By removing the fabric.properties file, building the release version
will now result in an gradle error if the fabric.properties file is
missing. This guarantees that the Crashlytics credentials have been
added to the release version.
> FAILURE: Build failed with an exception.
> Crashlytics Developer Tools error

The debug version will work as normal (because of
ext.enableCrashlytics = false) and will just log an error during runtime
which can be ignored.

> java.lang.IllegalArgumentException: Fabric could not be initialized,
  API key missing from AndroidManifest.xml